+++ src/standalone/Soil/soil_heat_parameterizations.jl
@@ -31,8 +31,8 @@ end
         τ::FT,
         ν::FT,
         θ_r::FT,
-        hydrology_earth_params::HEP
-    ) where {FT, HEP}
+        params::P
+    ) where {FT, P}
 Returns the source term (1/s) used for converting liquid water
 and ice into each other during phase changes. Note that
 there are unitless prefactors multiplying this term in the
@@ -49,11 +49,10 @@ function phase_change_source(
     τ::FT,
     ν::FT,
     θ_r::FT,
-    hydrology_earth_params::HEP,
-) where {FT, HEP}
+    params::P,
+) where {FT, P}
     # Extract parameter sets from their container
-    hydrology_cm = hydrology_earth_params.hydrology_cm
-    earth_param_set = hydrology_earth_params.earth_param_set
+    (; hydrology_cm, earth_param_set) = params
 
     _ρ_i = FT(LP.ρ_cloud_ice(earth_param_set))
     _ρ_l = FT(LP.ρ_cloud_liq(earth_param_set))
@@ -73,29 +72,6 @@ function phase_change_source(
 
     return (θ_l - θstar) / τ
 end
-
-"""
-    struct HydrologyEarthParameters{
-        HCM <: AbstractSoilHydrologyClosure,
-        EP <: ClimaLand.Parameters.AbstractLandParameters,
-    }
-
-A wrapper type around the hydrology closure model and land parameter
-structs. This is needed because of a type inference failure coming from
-ClimaCore when multiple structs and fields are broadcasted over.
-This struct circumvents that issue by wrapping the structs in a single
-type, that can be unpacked within the broadcasted function.
-
-See github.com/CliMA/ClimaCore.jl/issues/2065 for more information
-"""
-struct HydrologyEarthParameters{
-    HCM <: AbstractSoilHydrologyClosure,
-    EP <: ClimaLand.Parameters.AbstractLandParameters,
-}
-    hydrology_cm::HCM
-    earth_param_set::EP
-end
-Base.broadcastable(x::HydrologyEarthParameters) = tuple(x)
 
 """
     volumetric_heat_capacity(